zkDSL compiler: move free-variable analysis to ast.rs, and repair two spliced docs
The four free-variable functions take no lowering state and answer a question
about the syntax tree, so they sat in lower.rs only because their one caller is
the `for` desugaring. In ast.rs they are 162 lines the walker no longer carries;
lower.rs is 1,825 lines, from 3,124. The golden digests are unchanged, so the
bytecode is identical.

Two doc comments in FnLower were spliced by an earlier scripted edit. The
removed `alias` map's summary was left dangling above `slots`, and the removed
`phys` field's doc was grafted onto `cur_line`, truncated mid-clause at "if
either arm gives the cell a" so that cur_line appeared to document branch
restoration. What that fragment said is on `Slot` already, correctly. Three
comments elsewhere still named `phys`; they name `Slot::written` now.

The plan's reason for deferring M4 was that it depended on M3. That was wrong,
and the correction is that monomorphisation and loop desugaring depend
on scope, not on M3. A loop's capture set is its free variables filtered by what
each name is bound to, and `specialize` evaluates a constant argument, so
neither can run over an AST with no bindings.
