+++ crates/lean_compiler/src/lower.rs
@@ -266,10 +266,9 @@ struct FnLower<'a> {
     /// or a folded g-address, or take the scalar dst cell). `None` outside an
     /// inlined return.
     inline_stack_ret: Option<Vec<RetBind>>,
-    /// Deferred stack-cell copies/zeros ([`Alias`]), forwarded at use.
-    /// What is known about each frame cell. One entry per cell, replacing the two
-    /// side maps that answered "is this aliased" and "has this been written"
-    /// separately.
+    /// What is known about each frame cell ([`Slot`]). One entry per cell,
+    /// replacing the two side maps that answered "is this aliased" and "has this
+    /// been written" separately.
     slots: HashMap<Off, Slot>,
     /// The cells [`Self::stack_store`] has aliased, in the order it aliased
     /// them. A branch's join needs to know which cells it touched, and diffing
@@ -279,12 +278,6 @@ struct FnLower<'a> {
     /// order instead makes the join say what it produced rather than reconstruct
     /// it, which is the shape the middle IR needs.
     alias_journal: Vec<Off>,
-    /// Stack cells something already gives a real value to: an emitted
-    /// instruction's destination, a `BLAKE2s` output, or a hint destination. A
-    /// store into one of these cannot defer as an [`Alias`], because the store is
-    /// then the write-once equality assertion of `zkDSL.md` §Memory rather than an
-    /// assembly copy, and an alias would drop it. Accumulated monotonically, and
-    /// deliberately NOT restored across a branch: if either arm gives the cell a
     /// Source line of the statement being lowered, for diagnostics and for the
     /// pc-to-line table. Zero for a synthesized statement with no source.
     cur_line: u32,
@@ -1453,10 +1446,10 @@ impl FnLower<'_> {
             // One hinted value into one fresh cell, bound to `name`. The run form
             // has to materialize its destination first, since a deferred alias
             // there would take the hint's place. A cell minted here has no alias
-            // to displace, and more than that it can never acquire one: `phys` is
-            // read at exactly one site, `stack_store`, which is reached only
-            // through a `StackBuf` binding, so a scalar's cell is never a store
-            // target at all.
+            // to displace, and more than that it can never acquire one:
+            // `Slot::written` is read at exactly one site, `stack_store`, which is
+            // reached only through a `StackBuf` binding, so a scalar's cell is
+            // never a store target at all.
             StmtKind::LetHintWitness { name, stream } => {
                 let dst = self.fresh();
                 self.pending.push(Hint::Resolved(RHint::WitnessStack {
@@ -1714,168 +1707,6 @@ fn plus_k(lo: &Expr, hi: &Expr) -> Option<u128> {
             _ => None,
         },
         _ => None,
-    }
-}
-
-/// Collect variable references in `e` into `refs` (in source order).
-fn free_vars_expr(e: &Expr, refs: &mut Vec<String>) {
-    match e {
-        Expr::Var(v) => refs.push(v.clone()),
-        Expr::Add(a, b)
-        | Expr::Mul(a, b)
-        | Expr::Sub(a, b)
-        | Expr::Div(a, b)
-        | Expr::FieldDiv(a, b)
-        | Expr::Mod(a, b)
-        | Expr::Index(a, b)
-        | Expr::Pow(a, b) => {
-            free_vars_expr(a, refs);
-            free_vars_expr(b, refs);
-        }
-        Expr::Slice(a, lo, hi) => {
-            free_vars_expr(a, refs);
-            free_vars_expr(lo, refs);
-            free_vars_expr(hi, refs);
-        }
-        Expr::Call(_, args) | Expr::ListLit(args) => args.iter().for_each(|a| free_vars_expr(a, refs)),
-        Expr::HeapBufDyn(sz) | Expr::GenPow(sz) => free_vars_expr(sz, refs),
-        Expr::Lit(_) | Expr::Gen | Expr::GPow(_) | Expr::HeapBuf(_) | Expr::StackBuf(_) => {}
-    }
-}
-
-/// Collect references in `s` into `refs` and names it binds into `bound`.
-/// Every name the block binds, ignoring scope. [`free_vars_stmt`] deliberately
-/// does not answer this: its `bound` set is scoped, so an arm-local binding is
-/// discarded with the arm.
-fn binds_anywhere(body: &[Stmt], out: &mut std::collections::HashSet<String>) {
-    for s in body {
-        match &s.kind {
-            StmtKind::Let(n, _) | StmtKind::LetHintWitness { name: n, .. } => {
-                out.insert(n.clone());
-            }
-            StmtKind::LetTuple(ns, ..) | StmtKind::LetMatchRange { names: ns, .. } => ns.iter().for_each(|n| {
-                out.insert(n.clone());
-            }),
-            StmtKind::If { then, els, .. } => {
-                binds_anywhere(then, out);
-                binds_anywhere(els, out);
-            }
-            StmtKind::Match { cases, .. } => cases.iter().for_each(|c| binds_anywhere(c, out)),
-            StmtKind::For { var, body, .. } | StmtKind::Unroll { var, body, .. } => {
-                out.insert(var.clone());
-                binds_anywhere(body, out);
-            }
-            _ => {}
-        }
-    }
-}
-
-/// Free variables of a block whose bindings do NOT escape it: it sees everything
-/// bound so far, and anything it binds stays inside.
-fn scoped_vars(body: &[Stmt], refs: &mut Vec<String>, bound: &std::collections::HashSet<String>) {
-    let mut inner = bound.clone();
-    for s in body {
-        free_vars_stmt(s, refs, &mut inner);
-    }
-}
-
-fn free_vars_stmt(s: &Stmt, refs: &mut Vec<String>, bound: &mut std::collections::HashSet<String>) {
-    match &s.kind {
-        StmtKind::Let(n, e) => {
-            free_vars_expr(e, refs);
-            bound.insert(n.clone());
-        }
-        StmtKind::LetTuple(ns, _, args) => {
-            args.iter().for_each(|a| free_vars_expr(a, refs));
-            ns.iter().for_each(|n| {
-                bound.insert(n.clone());
-            });
-        }
-        StmtKind::AssertEq(a, b) | StmtKind::AssertNe(a, b) => {
-            free_vars_expr(a, refs);
-            free_vars_expr(b, refs);
-        }
-        StmtKind::AssertLt(e, bound) => {
-            free_vars_expr(e, refs);
-            if let LtBound::Runtime(b) = bound {
-                free_vars_expr(b, refs);
-            }
-        }
-        StmtKind::HintWitness { dest, .. } => free_vars_expr(dest, refs),
-        StmtKind::LetHintWitness { name, .. } => {
-            bound.insert(name.clone());
-        }
-        StmtKind::Print { value, .. } => free_vars_expr(value, refs),
-        StmtKind::If {
-            lhs,
-            rhs,
-            then,
-            els,
-            force_const,
-            ..
-        } => {
-            free_vars_expr(lhs, refs);
-            free_vars_expr(rhs, refs);
-            // An arm's bindings are local to it (`zkDSL.md` §Control flow), so each
-            // gets its own scope. Sharing one made a name rebound in ONE arm count
-            // as loop-local everywhere, so the OUTER binding the other arm reads
-            // was never captured and a legal program failed with `unbound
-            // variable`. Over-collecting into `refs` is harmless: a capture naming
-            // nothing in the enclosing scope is dropped.
-            // `lower_if` FOLDS a compile-time condition and runs the taken branch
-            // without `scoped`, so its bindings persist exactly like an `unroll`
-            // body's. Modelling that as scoped over-captured, and the loop's own
-            // self-call then read a name the folded arm had rebound to a
-            // `StackBuf`. Both sides literal is the syntactic half of that test.
-            if *force_const || matches!((lhs, rhs), (Expr::Lit(_), Expr::Lit(_))) {
-                then.iter().for_each(|s| free_vars_stmt(s, refs, bound));
-                els.iter().for_each(|s| free_vars_stmt(s, refs, bound));
-            } else {
-                scoped_vars(then, refs, bound);
-                scoped_vars(els, refs, bound);
-            }
-        }
-        StmtKind::Match { x, cases } => {
-            free_vars_expr(x, refs);
-            cases.iter().for_each(|c| scoped_vars(c, refs, bound));
-        }
-        StmtKind::LetMatchRange { names, x, arms } => {
-            free_vars_expr(x, refs);
-            arms.iter().for_each(|a| free_vars_expr(a, refs));
-            names.iter().for_each(|n| {
-                bound.insert(n.clone());
-            });
-        }
-        StmtKind::CallIfNe(a, b, _, args) => {
-            free_vars_expr(a, refs);
-            free_vars_expr(b, refs);
-            args.iter().for_each(|e| free_vars_expr(e, refs));
-        }
-        StmtKind::Call(_, args) => args.iter().for_each(|a| free_vars_expr(a, refs)),
-        StmtKind::Store(arr, idx, val) => {
-            free_vars_expr(arr, refs);
-            free_vars_expr(idx, refs);
-            free_vars_expr(val, refs);
-        }
-        StmtKind::Return(es) => es.iter().for_each(|e| free_vars_expr(e, refs)),
-        StmtKind::For { var, hi, body, .. } => {
-            if let ForBound::Runtime(b) = hi {
-                free_vars_expr(b, refs);
-            }
-            // A nested loop's body becomes its own function, so neither its
-            // counter nor its bindings exist out here. `unroll` below is the
-            // opposite: it replicates straight-line code into THIS scope, so its
-            // bindings really do persist and it keeps the shared set.
-            let mut inner = bound.clone();
-            inner.insert(var.clone());
-            body.iter().for_each(|s| free_vars_stmt(s, refs, &mut inner));
-        }
-        StmtKind::Unroll { var, lo, hi, body } => {
-            free_vars_expr(lo, refs);
-            free_vars_expr(hi, refs);
-            bound.insert(var.clone());
-            body.iter().for_each(|s| free_vars_stmt(s, refs, bound));
-        }
     }
 }
